* see-my-posts: (25 commits)
  Link photos to original pasted URL, not raw image URL.
  Change from 'keep this story private' to 'for TEACHER's eyes only.'
  Add typography styling to stories
  Remove unused data-width and data-heigh attributes on oembeds.
  Better label for add-message-form
  Fix broken landing page layout
  Switch to file SHA1 for asset cache busting.
  Make submitter field non-optional on stories.
  Check permissions correctly for delete/publish/hide actions.
  Add some TODO items.
  Fix templates test.
  Add login-to-post link on teacher-detail
  Fix teacher-detail template test.
  Fix login layout
  Fix broken layout
  Better access to embed-type in templates
  Oembeds working, layout broken
  Add <figure> wrapper around OEmbed markup.
  Apply oembed before markdown.
  Cache rendered story bodies.
  ...

Conflicts:
	sass/modules/_add-story.scss
	static/css/demo.css
	static/css/landing.css
	static/css/screen.css
	static/css/styleguide.css
	templates/_add_story_form.html

+++ knet/teachers/views.py
@@ -48,12 +48,19 @@ def teacher_detail(request, username):
         if not request.user.is_authenticated():
             return HttpResponseForbidden()
 
+        if any(
+            x in request.POST for x in
+            ['delete-story', 'publish-story', 'hide-story']
+            ) and (teacher_profile.user != request.user):
+            return HttpResponseForbidden()
+
         if 'delete-story' in request.POST:
             with transaction.atomic():
                 teacher_profile.stories.filter(
                     pk=request.POST['delete-story']).delete()
             messages.success(request, "Story deleted.")
             return _response(request, teacher)
+
         elif 'publish-story' in request.POST:
             with transaction.atomic():
                 story = get_or_none(
@@ -67,6 +74,7 @@ def teacher_detail(request, username):
                 else:
                     messages.error(request, "That story has been removed.")
             return _response(request, teacher, story, success=story is not None)
+
         elif 'hide-story' in request.POST:
             with transaction.atomic():
                 story = get_or_none(
